feat(users): tabbed Users screen + fold Batches/Categories into Master Data
Users screen now hosts both lists in shadcn Tabs:
- Tab 1: 'Users' — verified-user table + filters + Add/Edit/Delete
- Tab 2: 'Unverified Users' — pending-verification table

URL and tab state stay in sync: visiting /unverified-users opens the
screen on the second tab; switching tabs replaces the URL via
navigate(..., { replace: true }) so deep links and back-compat keep
working. Both routes render <UserList>; the standalone Unverified
Users sidebar entry is dropped.

Sidebar reorganization (continues the 'Master Data' grouping from
1a21175): Batches and Categories now also live under Master Data,
joining Users / Brands / Products / Product Offers / Product Catalog.
Sidebar root collapses from 14 leaves to 8 (Dashboard, Master Data,
Orders, Transactions, Ledger, Credit Notes, Reward Schemes, Payouts).

- Add shadcn tabs primitive
- Extract verified table into users-panel.tsx, unverified table into
  unverified-users-panel.tsx; new user-list.tsx wraps both in Tabs
- Delete now-empty unverified-users.tsx
- Drop unused UserX icon import from sidebar
- Route /unverified-users still resolves (renders UserList; initial
  tab derived from pathname)
